style: rename package from SatTrack to leosTrack. Add utils module
leos: low earth orbit satellites. Update imports. utils module contains functionality from old superclasses module
